Switch filename to get proper slug
