+++ website/src/pages/dev-tool-time.tsx
@@ -1,26 +1,116 @@
-import { graphql, PageProps } from 'gatsby'
+import { PageProps } from 'gatsby'
 import * as React from 'react'
-import { BLOG_TYPE_TO_INFO, BlogType } from '../components/blog/postTypes'
-import { PostsListPage } from '../components/blog/PostsListPage'
+import Layout from '../components/Layout'
+import { PostsList } from '../components/blog/PostsList'
+import { BlogType, BLOG_TYPE_TO_INFO, urlToPost } from '../components/blog/postTypes'
+import { BlogHeader } from '../components/blog/BlogHeader'
+import { FeaturedEpisode } from '../components/blog/FeaturedEpisode'
+import HubspotForm from '../components/HubspotForm'
 
-// TODOS:
-// -[x] Replace PostListPage w / subcomponents
-// -[x] Add featured video that is most recent
-// -[x] Add hubspot form
-// -[x] Add list of all guests, even without recaps
+// import Logo from '../../public/DevToolTimeLogo.svg'
+import upcomingProfile from '../../static/blog/lorna-jane-michell-profile.jpg'
 
-export const Page: React.FunctionComponent<PageProps<{ allMarkdownRemark: any }>> = props => (
-    <PostsListPage
-        blogInfo={BLOG_TYPE_TO_INFO[BlogType.DTT]}
-        posts={props.data.allMarkdownRemark.edges.filter((post: any) => post.node.frontmatter.published === true)}
-        location={props.location}
-    />
-)
+export const DttPage: React.FunctionComponent<PageProps<{ allMarkdownRemark: any }>> = props => {
+    const posts = props.data.allMarkdownRemark.edges.filter((post: any) => post.node.frontmatter.published === true)
+    const postDefaultProps = {
+        full: false,
+        className: 'posts-list__post card featured-episode',
+        headerClassName: 'card-header bg-white border-bottom-0 text-center',
+        titleClassName: 'posts-list__post-title',
+        titleLinkClassName: 'posts-list__post-title-link',
+        tag: 'li',
+        renderTitleAsLink: true,
+    }
+    return (
+        <Layout
+            location={props.location}
+            meta={{
+                title: 'Dev Tool Time',
+                description:
+                    'Sourcegraph engineers interview other devs in the community about their desk set up, favorite dev tools, productivity hacks, and more.',
+                image: 'https://about.sourcegraph.com/sourcegraph-og.png',
+            }}
+        >
+            <div className="container-lg">
+                <div className="dtt-header mt-4">
+                    <div className="dtt-brand">
+                        <p className="dtt-description mb-1">
+                            <BlogHeader title={'Dev Tool Time'} baseUrl={'/dtt'} />
+                            <p>
+                                Sourcegraph engineers interview other devs in the community about their desk set up,
+                                favorite dev tools, productivity hacks, and more. Episodes stream on{' '}
+                                <a href="https://www.twitch.tv/sourcegraph">Twitch</a> every other Wednesday at 11AM
+                                PT//2PM ET//8PM CEST.
+                            </p>
+                            <p>
+                                Hosted by <a href="https://twitter.com/TeejDeVries">TJ DeVries</a> and{' '}
+                                <a href="https://twitter.com/thorstenball">Thorsten Ball.</a>
+                            </p>
+                        </p>
+                    </div>
+                    {/* The mt class below is just to match the bulit-styling on the Blog Header */}
+                    {/* <img src={Logo} alt="Dev Tool Time Logo" className="dtt-logo mt-3" /> */}
+                </div>
+                <div className="dtt-upcoming card mt-3">
+                    <h2 className="card-header">Join us for our next episode!</h2>
+                    <div className="card-body upcoming-content">
+                        <div className="upcoming-description">
+                            <p className="upcoming-details">Lorna Jane Michell- Developer Advocate at Aiven</p>
+                            <p className="upcoming-date">December 8, 2021 | 11am PT / 2pm ET / 8pm CEST</p>
+                            <p>Fill in content.</p>
+                        </div>
+                        <div className="upcoming-profile-container">
+                            <img
+                                src={upcomingProfile}
+                                alt="Lorna Jane Michell profile image"
+                                className="upcoming-profile"
+                            />
+                        </div>
+                    </div>
+                </div>
+                <section className="py-5">
+                    {' '}
+                    <h2>Recent episodes:</h2>
+                    <FeaturedEpisode
+                        post={posts[0].node}
+                        url={urlToPost(posts[0].node, BLOG_TYPE_TO_INFO[BlogType.DTT])}
+                        key={posts[0].node.frontmatter.slug}
+                        {...postDefaultProps}
+                    />
+                    <div className="featured-episode-container">
+                        <FeaturedEpisode
+                            post={posts[1].node}
+                            url={urlToPost(posts[1].node, BLOG_TYPE_TO_INFO[BlogType.DTT])}
+                            key={posts[1].node.frontmatter.slug}
+                            {...postDefaultProps}
+                        />
+                        <div className="card p-4 hubspot-form">
+                            <h2>Sign up for reminders</h2>
+                            <HubspotForm
+                                portalId="2762526"
+                                formId="d59f5947-8f58-467b-8609-d095e9d8c9c7"
+                                campaignId="013t0000022mdYAAQ"
+                            />
+                        </div>
+                    </div>
+                </section>
+                <section>
+                    <h2>Other episodes:</h2>
+                    <p>
+                        You can view the full archive of Dev Tool Time episodes on{' '}
+                        <a href="https://www.youtube.com/playlist?list=PL6zLuuRVa1_iDEP4EicZ8972RgyccCRGF">Youtube</a>.
+                    </p>
+                    <PostsList blogInfo={BLOG_TYPE_TO_INFO[BlogType.DTT]} posts={posts.slice(2, -1)} />
+                </section>
+            </div>
+        </Layout>
+    )
+}
 
-export default Page
+export default DttPage
 
 export const pageQuery = graphql`
-    query DTTEpisodes {
+    query Dtt {
         allMarkdownRemark(
             filter: { frontmatter: { tags: { in: "dtt" } } }
             sort: { fields: [frontmatter___publishDate], order: DESC }
@@ -44,6 +134,7 @@ export const pageQuery = graphql`
                             description
                         }
                         style
+                        youtube
                     }
                     html
                     excerpt(pruneLength: 300)
